requirements.txt
streamlit
pytesseract
pillow
openai
python-dotenv
pandas
matplotlib